requirements.txt
# FastAPI et serveur
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Validation et modèles
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.0

# MQTT (optionnel pour communication hardware)
paho-mqtt==1.6.1

# Tests
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-html==4.1.1
httpx==0.25.2

# Utilitaires
python-dateutil==2.8.2

# Code quality (optionnel)
black==23.12.0
isort==5.13.2
pylint==3.0.3
mypy==1.7.1